requirements.txt
Flask==3.0.2
psycopg2==2.9.9
SQLAlchemy==2.0.27
PyDispatcher==2.0.7
pulsar-client==3.4.0
pulsar-client[avro]==3.4.0
python-dotenv==1.0.1
requests==2.28.2
aiopulsar-py==0.1.0a0
strawberry-graphql[fastapi]==0.156.3
sse-starlette==1.2.1
uvicorn[standard]==0.20.0
pydantic-settings==2.2.1
python-dotenv==1.0.1